Moved sn_flux_est to vcstools module
