+++ scripts/find_pulsar_in_obs.py
@@ -38,7 +38,7 @@ from astropy import units as u
 from vcstools import data_load
 from vcstools.pointing_utils import sex2deg, deg2sex, format_ra_dec
 
-import sn_flux_est as sfe
+import vcstools.sn_flux_est as sfe
 from mwa_pb import primary_beam
 from vcstools.metadb_utils import mwa_alt_az_za, get_common_obs_metadata,\
                              get_obs_array_phase, find_obsids_meta_pages,\
